requirements.txt
pandas
matplotlib
streamlit
yfinance
google-generativeai
fastapi
uvicorn
pydantic
Pillow
requests